use qulonglong construct to make the compiler cope with quintptr

+++ src/test/qsignaldumper.cpp
@@ -93,7 +93,7 @@ static void qSignalDumperCallback(QObject *caller, int method_index, void **argv
     str += objname.toLocal8Bit();
     if (!objname.isEmpty())
         str += ' ';
-    str += QByteArray::number(quintptr(caller), 16);
+    str += QByteArray::number(qulonglong(quintptr(caller)), 16);
 
     str += ") ";
     str += QTest::memberName(member);
@@ -111,7 +111,7 @@ static void qSignalDumperCallback(QObject *caller, int method_index, void **argv
                 str += '@';
 
             quintptr addr = quintptr(*reinterpret_cast<void **>(argv[i + 1]));
-            str.append(QByteArray::number(addr, 16));
+            str.append(QByteArray::number(qulonglong(addr), 16));
         } else if (typeId != QMetaType::Void) {
             str.append(arg)
                 .append('(')
@@ -149,7 +149,7 @@ static void qSignalDumperCallbackSlot(QObject *caller, int method_index, void **
     str += objname.toLocal8Bit();
     if (!objname.isEmpty())
         str += ' ';
-    str += QByteArray::number(quintptr(caller), 16);
+    str += QByteArray::number(qulonglong(quintptr(caller)), 16);
 
     str += ") ";
     str += member.signature();
